Separates the custom list detail view data source in a different class
